fix: send filter information to tessen for search events
